Update merge sort implementation and small changes
Slightly rebuild the way merge sort is implemented. First, there are comparisons in a loop between divided parts of the array. Results are stored in a temporary array and in the end, they are copied to an original array so that the visual part of merging can happen. Changed names of methods from the SortingElement class so that they follow camelCase rules.